requirements.txt
Django==2.2.1
django-templated-mail==1.1.1
djangorestframework==3.9.3
djoser==1.5.1
pytz==2019.1
sqlparse==0.3.0
